requirements.txt
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-eventhub>=11.0.0
azure-mgmt-cosmosdb>=9.0.0
azure-mgmt-web>=7.0.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
tabulate>=0.9.0
colorama>=0.4.6